Bump version number
So, I couldn't get the Design Manual to load the latest cf-icons code
from master, and as such, I was unable to use the new `-round` suffixes.
I realized that bower will only install the most recent _tagged
snapshot_ of a package. So this PR bumps the version to 0.2.0, and I'll
push the corresponding tag for it as soon as this PR gets merged.